pyproject.toml
[project]
name = "terraform-analysis-agent"
version = "0.1.0"
description = "MCP server for Terraform error debugging using RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.0",
    "openai>=1.12.0",
    "qdrant-client>=1.7.0",
    "tiktoken>=0.6.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.1.1",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
